% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{addvectors}
\alias{addvectors}
\title{Adds two vectors element-wise up to length n.}
\usage{
addvectors(x, y, n)
}
\arguments{
\item{x}{A reference to the first input vector (slice of f64)}

\item{y}{A reference to the second input vector (slice of f64)}

\item{n}{The number of elements to process}
}
\value{
The sum of the two input vectors, as a vector.
}
\description{
Adds two vectors element-wise up to length n.
}
