% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/022_calc_beta_params.R
\name{calc_beta_params}
\alias{calc_beta_params}
\title{Calculate Beta Distribution Parameters for a List of Events}
\usage{
calc_beta_params(mloss, sloss, expo)
}
\arguments{
\item{mloss}{mean loss by event}

\item{sloss}{sd loss by event}

\item{expo}{exposure by event}
}
\value{
A list containing beta distribution alpha and beta parameters, by event.
}
\description{
Calculates beta distribution parameters from by-event values of
mean, standard deviation and exposure.
}
\details{
Maybe I should change this one at some point so that it reads in an ELT, rather
than just separate variables. But for now it reads 3 separate variables.
}
\examples{
#
# create inputs
#
mloss=seq(100,300,100)
sloss=seq(100,300,100)
expo=rep(1000,3)
#
params=calc_beta_params(mloss,sloss,expo)
#
print(params$eventalpha)
print(params$eventbeta)
}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
