% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/806_hours_clause_compare_distance_routines.R
\name{compare_distance_routines}
\alias{compare_distance_routines}
\title{Function for testing that R and Rust give the same distances between events}
\usage{
compare_distance_routines(longylt)
}
\arguments{
\item{longylt}{Input YLT}
}
\value{
Writes mismatches to the screen.
Always stops.
}
\description{
This isn't actually used for anything, except doing this test.
But it's an important test. Getting R and Rust to agree was difficult.
}
\examples{
#
# 1 make longylt
#
cat("1. make longylt:\n")
year=c(1,1,1)
lon=c(-111.8817,-111.1963,-119.891)
lat=c(35.1067,31.4196,39.3254)
longylt=data.frame(year,lon,lat)
#
# 2 run
#
compare_distance_routines(longylt)
}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
