% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/900_countrydata_v16.R
\name{defineregion}
\alias{defineregion}
\title{Defines the 19 Global Regions from My BAMS Paper}
\usage{
defineregion(iregion)
}
\arguments{
\item{iregion}{Region index}
}
\value{
A list with two vectors, for longitudes and latitudes of corners of the region
}
\description{
Defines the 19 Global Regions from My BAMS Paper
}
\examples{
reg=defineregion(2)
cat("regx=",reg$regx,"\n")
cat("regy=",reg$regy,"\n")
}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
