% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/425_elt_add_nahu_cat.R
\name{elt_add_nahu_cat}
\alias{elt_add_nahu_cat}
\title{Adds a cat column to an elt, calculated from wind speed.}
\usage{
elt_add_nahu_cat(elt, units)
}
\arguments{
\item{elt}{A data frame containing the elt. Requires \code{wspd}.}

\item{units}{must be \code{knots}, or \code{mph}.}
}
\value{
A new elt with a cat column added, where cat runs from -1 to 5.
}
\description{
Takes an ELT, calculates NAHU cat from windspeed, and adds a column with cat.
The ELT has to contain windspeed to start with.
}
\details{
I've used the Wikipedia, and filled gaps with halves.
}
\examples{
#
# create an elt with just wind speed
#
wspd=seq(10,150,10)
elt1=data.frame(wspd)
#
# call
#
elt2=elt_add_nahu_cat(elt1,units="knots")
#
print(elt2)
}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
