% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/410_elt_diagnostics_aaeaal.R
\name{elt_diagnostics_aaeaal}
\alias{elt_diagnostics_aaeaal}
\title{Calculates AAE and AAL from an ELT}
\usage{
elt_diagnostics_aaeaal(elt)
}
\arguments{
\item{elt}{A data frame containing the ELT.
The ELT must contain \code{mrate} and \code{mloss} columns.
See the example for how to make the ELT.}
}
\value{
AAE and AAL.
EP curves are in a separate routine \code{epfromelt3}.
}
\description{
Calculates AAE and AAL from an ELT
}
\examples{
#
# create an elt
#
mrate=seq(0.1,0.3,0.1)
mloss=seq(100,300,100)
elt=data.frame(mrate,mloss)
#
op=elt_diagnostics_aaeaal(elt)
#
cat("AAE=",op$AAE,"\n")
cat("AAL=",op$AAL,"\n")
}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
