% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/411_elt_diagnostics_aaeaal_by_cat.R
\name{elt_diagnostics_aaeaal_by_cat}
\alias{elt_diagnostics_aaeaal_by_cat}
\title{Calculates AAE and AAL from an ELT, and by cat}
\usage{
elt_diagnostics_aaeaal_by_cat(elt)
}
\arguments{
\item{elt}{A data frame containing the ELT.
Must contain \code{mrate}, \code{mloss} and \code{cat} columns.
See the example for how to make the ELT.}
}
\value{
A list containing AAE and AAL, and then AAE and AAL by cat.
}
\description{
Calculates AAE and AAL from an ELT, and AAE and AAL by cat.
There can be any number of cats, labelled using integers.
So this routine is not only applicable to hurricane: the cat can indicate anything, for any peril.
}
\examples{
#
# create an elt with cat
#
mrate=seq(0.1,0.5,0.1)
mloss=seq(100,500,100)
cat=c(1,1,1,2,2)
elt=data.frame(mrate,mloss,cat)
#
op1=elt_diagnostics_aaeaal(elt)
op2=elt_diagnostics_aaeaal_by_cat(elt)
#
cat("AAE=",op1$AAE,"\n")
cat("AAL=",op1$AAL,"\n")
cat("AAEbycat=",op2$AAEbycat,"\n")
cat("AALbycat=",op2$AALbycat,"\n")
cat("AALbycatpc=",op2$AALbycatpc,"\n")
}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
