% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/420_elt_diagnostics_epcurves.R
\name{elt_diagnostics_epcurves}
\alias{elt_diagnostics_epcurves}
\title{Converts ELT rates into plotting positions for CEP, EEF and OEP}
\usage{
elt_diagnostics_epcurves(elt)
}
\arguments{
\item{elt}{needs \code{mrate} only}
}
\value{
A list containing plotting positions for CEP, EEF and OEP
}
\description{
Converts ELT rates into plotting positions for CEP, EEF and OEP
}
\details{
One could debate what plotting positions to use, as always.
This one uses Hazen. Perhaps Weibull would be better.
}
\examples{
#
# 1 make elt
#
nx=100
mrate=runif(nx)
elt=data.frame(mrate)
#
# 2 make the curves
#
op=elt_diagnostics_epcurves(elt)
#
# 3 add some loss data
#
loss=sort(rnorm(nx))
#
# 4 plot
#
old_par <- par(no.readonly = TRUE)
on.exit(par(old_par))
par(mfrow=c(2,2))
#
plot(loss,op$cep,main="CEP")
lines(loss,op$cep,col="red")
#
plot(loss,op$eef,main="EEF")
lines(loss,op$eef,col="red")
#
plot(loss,op$oep,main="OEP")
lines(loss,op$oep,col="red")
#


}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
