% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/422_elt_rate_adjustments_cat_2_event.R
\name{elt_rate_adjustments_cat_2_event}
\alias{elt_rate_adjustments_cat_2_event}
\title{Converts rate adjustments from cat to event, for events in an ELT}
\usage{
elt_rate_adjustments_cat_2_event(elt, rate_adjustments_by_cat)
}
\arguments{
\item{elt}{A data frame containing the ELT. Requires \code{cat}.}

\item{rate_adjustments_by_cat}{A list of two vectors containing the mean and sd adjustments by cat,
given by \code{adjustments_by_cat$mn}, \code{adjustments_by_cat$sd}}
}
\value{
A matrix with the adjustments by event
column 1 is the mean of the adjustments.
column 2 is the sd of the adjustments.
}
\description{
Converts rate adjustments by cat to rate adjustments by event.
So that you can specify rate adjustments by cat (which is easy to do),
and then run this routine to assign those adjustments to all the events
in an ELT.
}
\details{
Doesn't do anything except just copy the adjustments. No calculations.
Cat is a positive integer.
}
\examples{
#
# create an elt
#
cat=c(1,2,2,3,3,3) #only cat is required
elt=data.frame(cat)
#
# make the adjustments by cat
#
rate_adjustments_by_cat=list(mn=c(2,4,6),sd=c(0,0,0))
#
# call
#
rate_adjustments_by_event=elt_rate_adjustments_cat_2_event(elt,rate_adjustments_by_cat)
#
print(elt)
print(rate_adjustments_by_event)
}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
