% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/423_elt_rate_adjustments_regcat_2_event.R
\name{elt_rate_adjustments_regcat_2_event}
\alias{elt_rate_adjustments_regcat_2_event}
\title{Converts rate adjustments from region and cat to event, for events in an ELT}
\usage{
elt_rate_adjustments_regcat_2_event(elt, rate_adjustments_by_regcat)
}
\arguments{
\item{elt}{A data frame containing the ELT. Requires \code{cat} and \code{reg}}

\item{rate_adjustments_by_regcat}{A list of two matrices containing the mean and sd adjustments by reg-cat,
given by \code{adjustments_by_cat$mn}, \code{adjustments_by_cat$sd}}
}
\value{
A matrix with the adjustments by event
column 1 is the mean of the adjustments.
column 2 is the sd of the adjustments.
}
\description{
Converts rate adjustments by region and cat to rate adjustments by event.
So that you can specify rate adjustments by region and cat (which is easy to do),
and then run this routine to assign those adjustments to all the events
in an ELT.
}
\details{
Doesn't do anything except just copy the adjustments. No calculations.
Cat is a positive integer.
}
\examples{
#
# create an elt
#
cat_by_event=c(1,2,2,3,3,3)
reg_by_event=c(1,2,2,1,2,2)
elt=data.frame(cat=cat_by_event,region=reg_by_event)
#
# make the adjustments by reg-cat
# (2 regions, 3 cats)
#
mean=matrix(1,2,3)
sd=matrix(0,2,3)
rate_adjustments_by_regcat=list(mn=mean,sd=sd)
#
# call
#
rate_adjustments_by_event=elt_rate_adjustments_regcat_2_event(elt,rate_adjustments_by_regcat)
#
print(elt)
print(rate_adjustments_by_event)
}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
