% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/802_hours_clause_apply_part1.R
\name{hours_clause_apply_part1}
\alias{hours_clause_apply_part1}
\title{Function that loops through events in one year and applies an hours clause}
\usage{
hours_clause_apply_part1(
  hcdays,
  hckm,
  temploss,
  thisyearday,
  thisyearlon,
  thisyearlat,
  nevents_in_year,
  k
)
}
\arguments{
\item{hcdays}{Hours clause parameter, number of days}

\item{hckm}{Hours clause parameter, distance in km}

\item{temploss}{Not sure. Part of the algorithm.}

\item{thisyearday}{For events in this year, the days}

\item{thisyearlon}{For events in this year, the lons}

\item{thisyearlat}{For events in this year, the lats}

\item{nevents_in_year}{Number of events in this year}

\item{k}{The index of the hours clause to apply, of all possible hours clauses}
}
\value{
A loss and a counter
}
\description{
Quite hard to understand.
Considers hours clauses that start with the k'th event in the year,
and finish with the kk'th event, looping over kk.
Calculates the loss for each hours clause. The calling routine (one level
up) will loop over k, and look at all the losses, and choose the
hours clause that gives the largest loss.
}
\details{
To be called from \code{hours_clause}, which loops thru the whole YLT,
and which itself is called from \code{hours_clause_wrapper}, which manages the file i/o.
}
\examples{
#
# 1 set the hours clause settings
#
hcdays=15
hckm=200
#
# 2 make a single year of a ylt
#
cat("1. make longylt:\n")
nyearsinylt=8
thisyearloss	=seq(10,55,5)
thisyearday		=c(1	,2	,51	,101,151,201,251,300,350,360)
thisyearlat		=c(10	,11	,30	,30	,30	,30	,30	,30	,30	,50)	
thisyearlon		=c(10	,11	,30	,30	,30	,30	,30	,30	,30	,50)	
nevents_in_year1=10
#
# 3 apply the hours clause routine to position 1
#
cat("1. apply the hours clause calculator:\n")
output=hours_clause_apply_part1(hcdays,hckm,thisyearloss,
				thisyearday,thisyearlon,thisyearlat,nevents_in_year1,k=1)
#
# 4 look at the output
#
cat("losses=",output,"\n")
}
\references{
x
}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
