% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/805_hours_clause_write_settings.R
\name{hours_clause_write_settings}
\alias{hours_clause_write_settings}
\title{When using \code{hours_clause_wrapper}, writes out the settings to a file.}
\usage{
hours_clause_write_settings(
  ipfilename,
  settingsfilename,
  opfilename,
  time,
  distance
)
}
\arguments{
\item{ipfilename}{Input filename containing the input long YLT}

\item{settingsfilename}{Output filename for the settings output}

\item{opfilename}{Output filename containing the input long YLT}

\item{time}{Hours clause parameter, number of days}

\item{distance}{Hours clause parameter, distance in km}
}
\value{
Just writes to a file
}
\description{
Just for record keeping.
}
\details{
Just to provide a record of each analysis.
}
\examples{
#
# 1 make filenames
#
ipfilename=tempfile(fileext=".csv")
opfilename=tempfile(fileext=".csv")
settingsfilename=tempfile(fileext=".csv")
#
# 2 hc settings
#
time=1
distance=2
#
# 3 call the routine
#
hours_clause_write_settings(ipfilename,settingsfilename,opfilename,time,distance)
#
# 4 have a look at the file it produced
#
x=read.csv(settingsfilename)
print(x)


}
\references{
x
}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
