% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/027_input_checks.R
\name{input_checks}
\alias{input_checks}
\title{Checks the columns in an input data frame.}
\usage{
input_checks(df, names, location = FALSE)
}
\arguments{
\item{df}{data frame to check}

\item{names}{names to check}

\item{location}{print where it's being called from}
}
\value{
x
}
\description{
A utility that checks the columns in the input data frame versus a list of required names
}
\examples{
#
# make an elt
#
set.seed(1)
evid=c(1,2,3)
mrate=seq(0.1,0.3,0.1)
mloss=seq(100,300,100)
elt=data.frame(evid,mrate,mloss)
print(elt)
#
# and check if it contains the two columns
#
# one that passes:
input_checks(elt,c("evid","mrate","mloss"))
#
# and one that fails
#input_checks(elt,c("mrate","jim"))
}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
