% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/024_make_2_ep_by_sorting.R
\name{make_2_ep_by_sorting}
\alias{make_2_ep_by_sorting}
\title{Make 2 EP curves by sorting losses}
\usage{
make_2_ep_by_sorting(losses1, losses2, rps)
}
\arguments{
\item{losses1}{A vector of losses.}

\item{losses2}{A vector of losses.}

\item{rps}{The return periods at which to calculate the EPs}
}
\value{
A matrix with 2 sets of return levels in it (with 2 rows)
}
\description{
Calculates 2 EPs, from two vectors of losses, just by calling
\code{make_ep_by_sorting} twice.
This somewhat trivial routine is included because it is a very common
operation, to compare losses between two cases (two models, often
an original model and an adjusted model).
}
\examples{
#
# 1 create losses
#
cat("1. make losses:\n")
losses1=seq(1,1000,1)
losses2=seq(1001,2000,1)
#
# 2 set the return periods to look at
#
rps=c(1,2,3,4,5,10,20,50,100)
#
cat("2. calculate the EP:\n")
op=make_2_ep_by_sorting(losses1,losses2,rps)
#
print(op)

}
\seealso{
\itemize{
\item \code{make_ep_by_sorting} which is what this routine uses
}
}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
