% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003_make_elt.R
\name{make_elt}
\alias{make_elt}
\title{Make an ELT in memory}
\usage{
make_elt()
}
\value{
Nothing.
}
\description{
This code does nothing at all. But the example shows how to make an ELT in memory.
}
\details{
This code just exists as a way to present the example.
}
\examples{
#
# create an elt with 10 rows
#
set.seed(1)
#
# essentials
evid=seq(1000,1009,1) 
mrate=rep(0.1,10)
mloss=seq(100,1000,100)
#
# add an optional column
wspd=seq(50,59,1) 
#
# now make the elt
# the ordering of columns isn't important, but the following is standard
elt=data.frame(evid,mrate,mloss,wspd)
#
# and have a look at it, in different ways
print(elt)
head(elt,n=5)
}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
