% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/213_nahu_define_gmst_scenarios.R
\name{nahu_define_gmst_scenarios}
\alias{nahu_define_gmst_scenarios}
\title{Defines 4 GMST Scenarios}
\usage{
nahu_define_gmst_scenarios(verbose = FALSE)
}
\arguments{
\item{verbose}{logical}
}
\value{
An array with 4 GMST scenarios.
}
\description{
These are the 4 GMST scenarios for 2.6,...,8.5, by year,
derived from CMIP (from the JAMC paper Jewson(2021)).
From 1880 to 2100.
}
\examples{
gmst=nahu_define_gmst_scenarios()
print(head(gmst[,1:5]))

}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
