% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/214_nahu_define_k2020_zenodo_landfall_adj.R
\name{nahu_define_k2020_zenodo_landfall_adj}
\alias{nahu_define_k2020_zenodo_landfall_adj}
\title{Define Jewson landfall rates Scenarios from the BAMS paper}
\usage{
nahu_define_k2020_zenodo_landfall_adj()
}
\value{
An array with 19 landfall scenarios, described in the Jewson (2023) BAMS paper.
}
\description{
Multiplicative landfalling hurricane rate adjustments, mean and sd.
}
\details{
For the mean, a value of 1 is no change in the rate.
For the sd, value of 0 is no uncertainty in the rate change.
The distribution of rate changes is considered to be log-normal.
Index 1, from 1 to 19, specifies the region, as in the BAMS paper.
Index 2, specifies mean (1) or sd(2).
Index 3, specifies what set of categories. The 7 categories are -1,0,1,2,3,4,5 on the SSHWS.
}
\examples{
adjustments=nahu_define_k2020_zenodo_landfall_adj()
region=1
variance=1
cat=1
print(adjustments[1,1,1])

}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
