% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/215_nahu_k2020_rates_interpolation.R
\name{nahu_k2020_rates_interpolation}
\alias{nahu_k2020_rates_interpolation}
\title{Interpolates rates in time}
\usage{
nahu_k2020_rates_interpolation(
  rcp,
  baseyear1,
  baseyear2,
  targetyear,
  k2020settings,
  frequnc,
  quantile,
  randomseed,
  gmst,
  landfall,
  test = FALSE,
  verbose = FALSE,
  reflectinputs = FALSE
)
}
\arguments{
\item{rcp}{rcp}

\item{baseyear1}{baseyear1}

\item{baseyear2}{baseyear2}

\item{targetyear}{targetyear}

\item{k2020settings}{k2020settings}

\item{frequnc}{frequnc}

\item{quantile}{quantile}

\item{randomseed}{randomseed}

\item{gmst}{gmst}

\item{landfall}{landfall}

\item{test}{test}

\item{verbose}{verbose}

\item{reflectinputs}{reflectinputs}
}
\value{
Mean rate, sd rate, and GMST change
}
\description{
Longer description
}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
