% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/202_nahu_write_settings.R
\name{nahu_write_settings}
\alias{nahu_write_settings}
\title{Writes out settings to file for \code{nahu_ylt_surgery}}
\usage{
nahu_write_settings(
  ipfilename,
  settingsfilename,
  rcp,
  baseyear1,
  baseyear2,
  targetyear,
  k2020settings,
  frequnc,
  quantile,
  randomseed,
  manual,
  manualadjustments
)
}
\arguments{
\item{ipfilename}{ipfilename}

\item{settingsfilename}{settingsfilename}

\item{rcp}{rcp}

\item{baseyear1}{baseyear1}

\item{baseyear2}{baseyear2}

\item{targetyear}{targetyear}

\item{k2020settings}{k2020settings}

\item{frequnc}{frequnc}

\item{quantile}{quantile}

\item{randomseed}{randomseed}

\item{manual}{manual}

\item{manualadjustments}{manualadjustments}
}
\value{
Writes out a csv file
}
\description{
Writes out settings to file for \code{nahu_ylt_surgery}
}
\examples{
ipfilename="settings.csv"
settingsfilename=tempfile(fileext=".csv")
rcp=8
baseyear1=1900
baseyear2=2022
targetyear=2099
k2020settings="Linear and Landfall"
frequnc="Use distribution"
quantile=90
randomseed=0
manual=FALSE
manualadjustments=list(mn=1,sd=0)
#
nahu_write_settings(ipfilename,settingsfilename,rcp,baseyear1,baseyear2,targetyear,
		k2020settings,frequnc,quantile,randomseed,manual,manualadjustments)
}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
