% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_dist_haversine}
\alias{rust_dist_haversine}
\title{Function to calculate the Haversine distance between two points.
This matches the functionality of geosphere::distHaversine in R.}
\usage{
rust_dist_haversine(lon1, lat1, lon2, lat2)
}
\arguments{
\item{lon1}{Lon of first point.}

\item{lat1}{Lat of first point.}

\item{lon2}{Lon of second point.}

\item{lat2}{You guess.}
}
\value{
Returns distance in meters.

A message to the screen.
}
\description{
Function to calculate the Haversine distance between two points.
This matches the functionality of geosphere::distHaversine in R.
}
