% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_hello_world}
\alias{rust_hello_world}
\title{Return string \code{"Hello world!"} to R.
Just a test.}
\usage{
rust_hello_world()
}
\value{
A message to the screen.
}
\description{
Return string \code{"Hello world!"} to R.
Just a test.
}
