% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_hours_clause_apply_part2}
\alias{rust_hours_clause_apply_part2}
\title{Hours clause function
For a given start event, calculates all possible hours clauses, returns new losses AND a counter}
\usage{
rust_hours_clause_apply_part2(
  hcdays,
  hckm,
  temploss,
  thisyearday,
  thisyearlon,
  thisyearlat,
  nevents_in_year,
  k
)
}
\arguments{
\item{hcdays}{Definition of the hours clause in days}

\item{hckm}{Definition of the hours clause in km}

\item{temploss}{The losses by event for this year}

\item{thisyearday}{The days by event for this year}

\item{thisyearlon}{The lons by event for this year}

\item{thisyearlat}{The lats by event fo rthis year}

\item{nevents_in_year}{The number of events in the full year}

\item{k}{The starting event}
}
\value{
Returns a vector of the new losses
}
\description{
Hours clause function
For a given start event, calculates all possible hours clauses, returns new losses AND a counter
}
