% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_square}
\alias{rust_square}
\title{Square an integer.
Just an example.}
\usage{
rust_square(x)
}
\arguments{
\item{x}{An integer.}
}
\value{
The square of the input.
}
\description{
Square an integer.
Just an example.
}
