% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002_stop_intentionally.R
\name{stopi}
\alias{stopi}
\title{Stops code intentionally, with message that says that}
\usage{
stopi(location = FALSE)
}
\arguments{
\item{location}{An optional string giving the routine being called from, which is then
included in the message.}
}
\value{
A message to the screen.
}
\description{
Stops execution and returns a message saying the stop was intentional
}
\details{
Not much more to say about this really.
}
\examples{
# commented out to pass CRAN tests, which don't allow execution to halt 
# stopi("example")

}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
