% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/026_whatreg.R
\name{whatreg}
\alias{whatreg}
\title{Converts US state and country region abbreviations into numbers from 1 to 7.}
\usage{
whatreg(rg)
}
\arguments{
\item{rg}{Region abbreviations}
}
\value{
The region from 1 to 7
}
\description{
This is based on one particular method for assigning 7 regions to a NAHU model.
}
\details{
Uses the following regions:

region 1=c("TX","LA","MS","AL")

region 2=c("FL")

region 3=c("GA","SC","NC","VA")

region 4=c("MD","DE","RI","MA","ME","NY","NJ")

region 5=c("Mexico")

region 6=c("Panama","Costa Rica","Nicaragua","Honduras","El Salvador","Gautemala","Belize")

region 7=undefined
}
\examples{
reg="Mexico"
message("region input name = ",reg)
message("region output #   = ",whatreg(reg))
}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
