% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/554_ylt_writeNresults2csv.R
\name{writeNresults2csv}
\alias{writeNresults2csv}
\title{Writes results from N YLTs to csv format.}
\usage{
writeNresults2csv(
  nyearsinylt,
  mincat,
  maxcat,
  nrp,
  nylt,
  maxnumberevents,
  bsaepsd,
  bsoepsd,
  aal,
  pcaal,
  aalbc,
  pcaalbc,
  aae,
  pcaae,
  aaebc,
  pcaaebc,
  nbc,
  probs,
  rps,
  aep,
  daep,
  oep,
  doep,
  lossthresholds,
  thresholdcounts,
  justone = FALSE,
  iylt
)
}
\arguments{
\item{nyearsinylt}{number of years}

\item{mincat}{minimum cat}

\item{maxcat}{maximum cat}

\item{nrp}{number of return periods}

\item{nylt}{number of ylts}

\item{maxnumberevents}{related to the threshold counting}

\item{bsaepsd}{bootstraps sds for AEP}

\item{bsoepsd}{bootstraps sds for OEP}

\item{aal}{AAL results}

\item{pcaal}{AAL as percentages}

\item{aalbc}{AAL results}

\item{pcaalbc}{AAL as percentages by cat}

\item{aae}{AAE results}

\item{pcaae}{AAE as percentages}

\item{aaebc}{AAE results by cat}

\item{pcaaebc}{AAE results as percentage}

\item{nbc}{Number by cat results}

\item{probs}{return probabilities being used}

\item{rps}{return levels being used}

\item{aep}{AEP results}

\item{daep}{AEP changes}

\item{oep}{OEP results}

\item{doep}{OEP changes}

\item{lossthresholds}{For the threshold counting}

\item{thresholdcounts}{threshold counting results}

\item{justone}{logical if there's just one run to be processed}

\item{iylt}{if there's just one, what is the i}
}
\value{
CSV output, for screen or file.
}
\description{
Writes various results from N YLTs to csv.
Can be used with \code{sink()} to make a nicely formatted csv file.
The result are:
AAE and AAL by cat, AEP, OEP and threshold counting.
}
\examples{
#' Uses the example for routine 201, \code{nahu_ylt_surgery}.
}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
