% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/540_ylt_add_region_2_longylt.R
\name{ylt_add_region_2_longylt}
\alias{ylt_add_region_2_longylt}
\title{Adds a region column to a longylt, based on the lat-lons already in the longylt}
\usage{
ylt_add_region_2_longylt(longylt)
}
\arguments{
\item{longylt}{A data frame containing the long YLT. Must contain
\code{year}, \code{lflat}, \code{lflon}.}
}
\value{
A new ylt with a region column added
}
\description{
Calculates region from lat lon
(where the regions relate to the Jewson (2023) BAMS paper).
}
\examples{
#
# 1 create a YLT
#
cat("1. longylt:\n")
year=c(1,2)
lflat=c(40,10)
lflon=c(-40,-40)
loss=c(10,20)
longylt=data.frame(year,loss,lflat,lflon)
print(longylt)
#
op=ylt_add_region_2_longylt(longylt)
print(op)
}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
