% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/537_ylt_check_for_2_events_on_same_day.R
\name{ylt_check_for_2_events_on_same_day}
\alias{ylt_check_for_2_events_on_same_day}
\title{Checks a YLT for two events on the same day}
\usage{
ylt_check_for_2_events_on_same_day(longylt)
}
\arguments{
\item{longylt}{The input longylt, which needs a \code{day} column.}
}
\value{
Screen output listing when there are two events on the same day
}
\description{
The assumption is that two events on the same day is bad, so if it
finds two events on the same day, it stops running, and gives a message
saying what day they are on.
}
\examples{
#
# 1 create a long YLT with cat
#
cat("1. make longylt:\n")
year=c(1,1,1,2,2)
day=c(1,2,3,4,5)
loss=rep(0,5) #needs an mloss column for long2short
longylt=data.frame(year,loss,day)
#
# 2 run the check
#
cat("4. run the diagnostics:\n")
op=ylt_check_for_2_events_on_same_day(longylt)
}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
