% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/538_ylt_check_for_same_events_in_one_year.R
\name{ylt_check_for_same_events_in_one_year}
\alias{ylt_check_for_same_events_in_one_year}
\title{Checks a YLT for the same event occurring twice in one year}
\usage{
ylt_check_for_same_events_in_one_year(longylt)
}
\arguments{
\item{longylt}{The input long, which needs a \code{evid} column.}
}
\value{
Screen output listing when an event occurs twice in the same year
}
\description{
The assumption is that this is undesirable, and so if it happens
then the code stops and reports back when it happened.
}
\examples{
#
# 1 create a long YLT with evid
#
cat("1. make longylt:\n")
year=c(1,1,1,2,2)
loss=rep(0,5) 
evid=c(1,2,3,4,5)
longylt=data.frame(year,loss,evid)
#
# 2 run the check
#
cat("4. run the diagnostics:\n")
op=ylt_check_for_same_events_in_one_year(longylt)
}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
