% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/580_ylt_diagnostics_aaeaal.R
\name{ylt_diagnostics_aaeaal}
\alias{ylt_diagnostics_aaeaal}
\title{Long YLT Diagnostics}
\usage{
ylt_diagnostics_aaeaal(longylt, nyearsinylt = 0)
}
\arguments{
\item{longylt}{A data frame containing the long YLT}

\item{nyearsinylt}{Sometimes, by chance, the long YLT might not have any
events in year 100, even
though technically it's supposed to be 100 years long. So it's good to specify
the number of actual years (in this example, 100). Then the short YLT will have
100 years in it. The last year just won't have any events in it. If you just leave
this to the default value of 0, then the short YLT will just stop at the last year
in the long YLT.}
}
\value{
AAE, AAL, SAE and SAL.
}
\description{
Long YLT diagnostics,
Works by creating a shortylt using \code{ylt_long2short}, and then using \code{ylt_diagnostics_aaeaal}.
A bit stupid because the output contains blank weighted values, because \code{ylt_long2short},
returns weighted values, but there's no way in this case to set the weights.
Oh well, never mind. It's fine.
You don't really need this one. You can always use \code{ylt_long2short} yourself,
and then use \code{ylt_diagnostics_aaeaal}.
So this is a good candidate for deletion tbh.
}
\examples{
#
# 1 create an ELT
#
evid=c(1,2,3)
mrate=c(1,1,1)
mloss=c(10,20,30)
elt=data.frame(evid,mrate,mloss)
#
# 2 create a YLT
#
longylt=yltsim(1000,elt)
#
# 3 call diagnostics
#
op=ylt_diagnostics_aaeaal(longylt)
#
cat(" AAE=",op$AAE,"\n")
cat(" AAL=",op$AAL,"\n")
cat(" SAE=",op$SAE,"\n")
cat(" SAL=",op$SAL,"\n")
}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
