% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/585_ylt_diagnostics_aaeaal_by_cat_change_printf.R
\name{ylt_diagnostics_aaeaal_by_cat_change_printf}
\alias{ylt_diagnostics_aaeaal_by_cat_change_printf}
\title{Prints changes in AAE and AAL from a two YLTs to the screen}
\usage{
ylt_diagnostics_aaeaal_by_cat_change_printf(
  longylt1,
  longylt2,
  mincat,
  maxcat,
  nyearsinylt = 0,
  verbose = FALSE
)
}
\arguments{
\item{longylt1}{A data frame containing YLT1}

\item{longylt2}{A data frame containing YLT2}

\item{mincat}{The minimum cat that is counted}

\item{maxcat}{The maximum cat that is counted}

\item{nyearsinylt}{Number of years in the YLT}

\item{verbose}{Logical}
}
\value{
Prints changes in AAE, AAL to the screen
}
\description{
Only works for the case where there are \emph{exactly} 5 cats in both YLTs.
Which is a bit limiting. Could improve that I guess.
}
\examples{
#
# 1 make YLT1
#
cat("1. make ylt1:\n")
year=c(1,1,1,2,2,3,3,3)
loss=seq(10,80,10)
cat=c(1,1,1,2,2,3,4,5)
longylt1=data.frame(year,loss,cat)
#
# 2 make YLT2
#
cat("1. make ylt2:\n")
year=c(1,1,2,2,2,2,3,3)
loss=seq(20,90,10)
cat=c(1,1,1,2,2,3,4,5)
longylt2=data.frame(year,loss,cat)
#
# look at changes
#
cat("calling ylt_change_diagnostics_aaeaal_by_cat_print:\n")
ylt_diagnostics_aaeaal_by_cat_change_printf(longylt1,longylt2,mincat=1,maxcat=5)
}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
