% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/584_ylt_diagnostics_aaeaal_by_cat_printf.R
\name{ylt_diagnostics_aaeaal_by_cat_printf}
\alias{ylt_diagnostics_aaeaal_by_cat_printf}
\title{Prints AAE and AAL from a YLT to the screen}
\usage{
ylt_diagnostics_aaeaal_by_cat_printf(
  longylt,
  nyearsinylt = 0,
  mincat,
  maxcat,
  verbose = FALSE
)
}
\arguments{
\item{longylt}{The YLT}

\item{nyearsinylt}{The number of years in the YLT.}

\item{mincat}{The minimum cat that is counted}

\item{maxcat}{The maximum cat that is counted}

\item{verbose}{Logical}
}
\value{
Prints changes in AAE, AAL to the screen
}
\description{
Just a way to print the results from \code{ylt_diagnostics_aaeaal_by_catf}
to the screen in a pretty way.
}
\examples{
#
# 1 make YLT
#
cat("1. make ylt1:\n")
year=c(1,1,1,2,2,3,3,3)
loss=seq(10,80,10)
cat=c(1,1,1,2,2,3,4,5)
longylt=data.frame(year,loss,cat)
#
# 2 look 
#
cat("2. print the diagnostics\n")
op=ylt_diagnostics_aaeaal_by_cat_printf(longylt,mincat=1,maxcat=5)
}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
