% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/582_ylt_diagnostics_aaeaal_change_print.R
\name{ylt_diagnostics_aaeaal_change_printf}
\alias{ylt_diagnostics_aaeaal_change_printf}
\title{Prints changes in AAE and AAL from a two longylts to the screen}
\usage{
ylt_diagnostics_aaeaal_change_printf(longylt1, longylt2, nyearsinylt = 0)
}
\arguments{
\item{longylt1}{A data frame containing longylt1}

\item{longylt2}{A data frame containing longylt2}

\item{nyearsinylt}{nyearsinylt}
}
\value{
Prints changes in AAE, AAL to the screen
}
\description{
Because they are longylts, requires the number of years.
}
\examples{
#
# 1 create an ELT
#
evid=c(1,2,3)
mrate=c(1,1,1)
mloss=c(10,20,30)
elt=data.frame(evid,mrate,mloss)
#
# 2 create YLTs
#
longylt1=yltsim(1000,elt)
longylt2=yltsim(1000,elt)
#
# 3 call diagnostics
#
ylt_diagnostics_aaeaal_change_printf(longylt1,longylt2)
#
}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
