% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/581_ylt_diagnostics_aaeaal_printf.R
\name{ylt_diagnostics_aaeaal_printf}
\alias{ylt_diagnostics_aaeaal_printf}
\title{Prints AAE and AAL from a Long YLT to the screen}
\usage{
ylt_diagnostics_aaeaal_printf(longylt, nyearsinylt = 0)
}
\arguments{
\item{longylt}{A data frame containing the long YLT}

\item{nyearsinylt}{Sometimes, by chance, the long YLT might not have any
events in year 100, even
though technically it's supposed to be 100 years long. So it's good to specify
the number of actual years (in this example, 100). Then the short YLT will have
100 years in it. The last year just won't have any events in it. If you just leave
this to the default value of 0, then the short YLT will just stop at the last year
in the long YLT.}
}
\value{
Prints AAE, AAL to the screen
}
\description{
Because it's a longylt, you have to specify the \code{nyearsinylt}.
If you have a full ylt, better to use \code{ylt_diagnostics_aaeaal_print},
because then you don't have to specify that.
}
\examples{
#
# 1 create an ELT
#
evid=c(1,2,3)
mrate=c(1,1,1)
mloss=c(10,20,30)
elt=data.frame(evid,mrate,mloss)
#
# 2 create a YLT
#
longylt=yltsim(1000,elt)
#
# 3 call diagnostics
#
ylt_diagnostics_aaeaal_printf(longylt)
#
}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
