% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/552_ylt_write2results2csv.R
\name{ylt_write2results2csv}
\alias{ylt_write2results2csv}
\title{Writes results from 2 YLTs to csv format.}
\usage{
ylt_write2results2csv(longylt1, longylt2, rps)
}
\arguments{
\item{longylt1}{The first YLT}

\item{longylt2}{The first YLT}

\item{rps}{The return periods you want to calculate}
}
\value{
A csv file.
}
\description{
Writes various results from 2 YLTs to a csv format.
Not to a csv file...but to standard output that could be sent to a csv file
using the \code{sink()} function.
The result are:
AAE and AAL by cat, and AEP.
}
\examples{
#
# 1 create YLT1
#
year=c(1,1,2,2,3)
cat=c(1,2,3,4,5)
loss=seq(10,50,10)
longylt1=data.frame(year,cat,loss)
#
# 2 create YLT2
#
year=c(1,1,2,2,3)
cat=c(1,2,3,4,5)
loss=seq(110,150,10)
longylt2=data.frame(year,cat,loss)
#
# 3 do it
#
rps=c(2,5,10,50,100,130,200,250,500,1000,5000)
ylt_write2results2csv(longylt1,longylt2,rps)


}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
